this updates to the latest gvrf.
this updates to the latest gvrf.  also modifies the note4.xml to use
monoscopic and 2xmsaa.

problem:  most of the time i see everything in green.  as if the texture
isn't loaded.  we sometimes see this when multiple objects are using the
same FutureTexture.  you can perhaps use the (deprecated) blocking
loadTexture() instead.